+++ src/main/java/ie/futurecollars/invoicing/controller/invoice/InvoiceController.java
@@ -22,7 +22,7 @@ public class InvoiceController implements InvoiceApi {
   }
 
   @Override
-  public int add(@RequestBody Invoice invoice) {
+  public long add(@RequestBody Invoice invoice) {
     return invoiceService.save(invoice);
   }
