+++ src/main/java/ie/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -20,19 +20,19 @@ public class FileBasedDatabase implements Database {
   private final JsonService jsonService;
 
   @Override
-  public int save(Invoice invoice) {
+  public long save(Invoice invoice) {
     try {
       invoice.setId(idProvider.getNextIdAndIncrement());
       filesService.appendLineToFile(databasePath, jsonService.toJson(invoice));
 
-      return invoice.getId();
+      return (int) invoice.getId();
     } catch (IOException ex) {
       throw new RuntimeException("Database failed to save invoice", ex);
     }
   }
 
   @Override
-  public Optional<Invoice> getById(int id) {
+  public Optional<Invoice> getById(long id) {
     try {
       return filesService.readAllLines(databasePath)
           .stream()
@@ -57,7 +57,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+  public Optional<Invoice> update(long id, Invoice updatedInvoice) {
     try {
       List<String> allInvoices = filesService.readAllLines(databasePath);
       var invoicesWithoutInvoiceWithGivenId = allInvoices
@@ -79,7 +79,7 @@ public class FileBasedDatabase implements Database {
   }
 
   @Override
-  public Optional<Invoice> delete(int id) {
+  public Optional<Invoice> delete(long id) {
     try {
       var allInvoices = filesService.readAllLines(databasePath);
 
@@ -98,7 +98,7 @@ public class FileBasedDatabase implements Database {
     }
   }
 
-  private boolean containsId(String line, int id) {
+  private boolean containsId(String line, long id) {
     return line.contains("\"id\":" + id + ",");
   }
 }
